pyproject.toml
[tool.poetry]
name = "botsreustacio"
version = "0.1.0"
description = ""
authors = ["Rich <richelton14@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
python-dotenv = "^0.18.0"
gTTS = "^2.2.3"
PyNaCl = "^1.4.0"
databases = {extras = ["sqlite"], version = "^0.4.3"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
